bot/main.py
from telegram import Update
from telegram.ext import Application, CommandHandler, MessageHandler, CallbackContext, filters
from handlers import welcome_new_user,  delete_left_message
from config import BOT_TOKEN


def main():
    app = Application.builder().token(BOT_TOKEN).build()

    # Handlerlar
    app.add_handler(MessageHandler(filters.StatusUpdate.NEW_CHAT_MEMBERS, welcome_new_user))

    app.add_handler(MessageHandler(filters.StatusUpdate.LEFT_CHAT_MEMBER, delete_left_message))


    # Botni ishga tushirish
    app.run_polling()


if __name__ == "__main__":
    main()