requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==21.8
sniffio==1.3.1
